requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1

# AI
anthropic==0.44.0

# Dashboard
streamlit==1.41.1
plotly==5.24.1

# Data processing
pandas==2.2.3
numpy==2.2.1

# HTTP clients (API integrations)
httpx==0.28.1
wyze-sdk==2.2.0

# Database
aiosqlite==0.20.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0
